pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
package-mode = true
authors = ["lihuan <343397495@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sanic = "^24.6.0"
sqlalchemy = "^2.0.31"
aiomysql = "^0.2.0"
requests = "^2.32.3"
redis = "^5.0.7"
pytest = "^8.3.2"
numpy = "1.26.4"
pyyaml = "^6.0.1"
duckdb = "^1.0.0"
openpyxl = "^3.1.5"
seaborn = "^0.13.2"
langchain = "^0.2.14"
pdfplumber = "^0.11.3"
python-docx = "^1.1.2"
ollama = "^0.3.1"
langchain-community = "^0.2.12"
python-dotenv = "^1.0.1"
pandasai = "^2.2.15"
minio = "^7.2.8"
pyjwt = "^2.9.0"
selenium = "^4.26.0"
bs4 = "^0.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# 格式化代码插件
[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''/(\.git|\.hg|\.mypy_cache|\.tox|\.venv|_build|buck-out|build|dist)/'''